Refactor namespaces and documentation
- Moved classes from Reflection\OFXAttributes folder to Attributes folder;
- Removed ver102 namespace from GlobalUsings;
- Moved comments to remarks xml documentation in OFXSpecification class;
