+++ src/OFXSDK.sln
@@ -13,9 +13,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Misc", "Misc", "{BACD69CE-6
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 		..\LICENSE = ..\LICENSE
+		..\README.md = ..\README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OFX.SDK.Tests", "OFX.SDK.Tests\OFX.SDK.Tests.csproj", "{EB0DB030-9ABB-4831-AFF9-6541459AF025}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OFX.SDK.Tests", "OFX.SDK.Tests\OFX.SDK.Tests.csproj", "{EB0DB030-9ABB-4831-AFF9-6541459AF025}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
